Take into account Excon response time
We were measuring only the Excon request time, which if I understand correctly
is the time taken to send the request to the server, it does not include the
time it takes to read the response back from the socket.

This could explain why we've seen very small request times since the request
happens with a persistent connection.

It also fixes the issue of adding integers and floats together.